Clean-up dispatcher to run only as daemon & fix sanitizer inconsistency with dispatcher
sanitizer will now respect blocked clusters the same way dispatcher does
