+++ cmd/prow-job-dispatcher/main_test.go
@@ -19,74 +19,6 @@ import (
 	"github.com/openshift/ci-tools/pkg/testhelper"
 )
 
-func TestValidate(t *testing.T) {
-	testCases := []struct {
-		name     string
-		given    *options
-		expected error
-	}{
-		{
-			name:     "empty options",
-			given:    &options{},
-			expected: fmt.Errorf("mandatory argument --prow-jobs-dir wasn't set"),
-		},
-		{
-			name: "set only prow-jobs-dir",
-			given: &options{
-				prowJobConfigDir: "--prow-jobs-dir",
-			},
-			expected: fmt.Errorf("mandatory argument --config-path wasn't set"),
-		},
-		{
-			name: "OK if no username and no possword",
-			given: &options{
-				prowJobConfigDir:     "prow-jobs-dir",
-				configPath:           "some-path",
-				prometheusDaysBefore: 1,
-			},
-		},
-		{
-			name: "prometheus username is set while password not",
-			given: &options{
-				prowJobConfigDir:     "prow-jobs-dir",
-				configPath:           "some-path",
-				prometheusDaysBefore: 1,
-				PrometheusOptions: dispatcher.PrometheusOptions{
-					PrometheusUsername: "user",
-				},
-			},
-			expected: fmt.Errorf("--prometheus-username and --prometheus-password-path must be specified together"),
-		},
-		{
-			name: "prometheus password path is set while username not",
-			given: &options{
-				prowJobConfigDir:     "prow-jobs-dir",
-				configPath:           "some-path",
-				prometheusDaysBefore: 1,
-				PrometheusOptions: dispatcher.PrometheusOptions{
-					PrometheusPasswordPath: "some-path",
-				},
-			},
-			expected: fmt.Errorf("--prometheus-username and --prometheus-password-path must be specified together"),
-		},
-		{
-			name: "prometheus days before cannot be 16",
-			given: &options{
-				prowJobConfigDir:     "prow-jobs-dir",
-				configPath:           "some-path",
-				prometheusDaysBefore: 16,
-			},
-			expected: fmt.Errorf("--prometheus-days-before must be between 1 and 15"),
-		},
-	}
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			actual := tc.given.validate()
-			equalError(t, tc.expected, actual)
-		})
-	}
-}
-
 var (
 	c = dispatcher.Config{
 		DetermineE2EByJob: true,
@@ -467,97 +399,6 @@ func TestRemoveDisabledClusters(t *testing.T) {
 				}
 			}
 
-		})
-	}
-}
-
-func TestDetermineCluster(t *testing.T) {
-	type fields struct {
-		blocked sets.Set[string]
-	}
-	type args struct {
-		cluster           string
-		determinedCluster string
-		defaultCluster    string
-		canBeRelocated    bool
-	}
-	tests := []struct {
-		name   string
-		fields fields
-		args   args
-		want   string
-	}{
-		{
-			name: "relocate to cluster for a test group",
-			fields: fields{
-				blocked: sets.New[string](),
-			},
-			args: args{
-				cluster:           "build01",
-				determinedCluster: "build02",
-				defaultCluster:    "build03",
-				canBeRelocated:    true,
-			},
-			want: "build01",
-		},
-		{
-			name: "can't relocate to cluster for a test group",
-			fields: fields{
-				blocked: sets.New[string](),
-			},
-			args: args{
-				cluster:           "build01",
-				determinedCluster: "build02",
-				defaultCluster:    "build03",
-				canBeRelocated:    false,
-			},
-			want: "build02",
-		},
-		{
-			name: "both clusters are blocked, relocate to default",
-			fields: fields{
-				blocked: sets.New[string]("build01", "build02"),
-			},
-			args: args{
-				cluster:           "build01",
-				determinedCluster: "build02",
-				defaultCluster:    "build03",
-				canBeRelocated:    false,
-			},
-			want: "build03",
-		},
-		{
-			name: "determined is blocked, relocate to a group cluster despite canBeRelocated=false",
-			fields: fields{
-				blocked: sets.New[string]("build02"),
-			},
-			args: args{
-				cluster:           "build01",
-				determinedCluster: "build02",
-				defaultCluster:    "build03",
-				canBeRelocated:    false,
-			},
-			want: "build01",
-		},
-		{
-			name: "group cluster is blocked, use determined cluster",
-			fields: fields{
-				blocked: sets.New[string]("build01"),
-			},
-			args: args{
-				cluster:           "build01",
-				determinedCluster: "build02",
-				defaultCluster:    "build03",
-				canBeRelocated:    false,
-			},
-			want: "build02",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := determineCluster(tt.args.cluster, tt.args.determinedCluster, tt.args.defaultCluster, tt.args.canBeRelocated, tt.fields.blocked); got != tt.want {
-				t.Errorf("clusterVolume.determineCluster() = %v, want %v", got, tt.want)
-			}
 		})
 	}
 }
