+++ cmd/sanitize-prow-jobs/main.go
@@ -12,8 +12,9 @@ import (
 )
 
 type options struct {
-	prowJobConfigDir string
-	configPath       string
+	prowJobConfigDir  string
+	configPath        string
+	clusterConfigPath string
 
 	help bool
 }
@@ -23,6 +24,7 @@ func bindOptions(flag *flag.FlagSet) *options {
 
 	flag.StringVar(&opt.prowJobConfigDir, "prow-jobs-dir", "", "Path to a root of directory structure with Prow job config files (ci-operator/jobs in openshift/release)")
 	flag.StringVar(&opt.configPath, "config-path", "", "Path to the config file (core-services/sanitize-prow-jobs/_config.yaml in openshift/release)")
+	flag.StringVar(&opt.clusterConfigPath, "cluster-config-path", "core-services/sanitize-prow-jobs/_clusters.yaml", "Path to the config file (core-services/sanitize-prow-jobs/_clusters.yaml in openshift/release)")
 	flag.BoolVar(&opt.help, "h", false, "Show help for ci-operator-prowgen")
 
 	return opt
@@ -51,6 +53,10 @@ func main() {
 	if err != nil {
 		logrus.WithError(err).Fatalf("Failed to load config from %q", opt.configPath)
 	}
+	_, blocked, err := sanitizer.LoadClusterConfig(opt.clusterConfigPath)
+	if err != nil {
+		logrus.WithError(err).Fatalf("Failed to load cluster config from %q", opt.configPath)
+	}
 	if err := config.Validate(); err != nil {
 		logrus.WithError(err).Fatal("Failed to validate the config")
 	}
@@ -60,7 +66,7 @@ func main() {
 	}
 	for _, subDir := range args {
 		subDir = filepath.Join(opt.prowJobConfigDir, subDir)
-		if err := sanitizer.DeterminizeJobs(subDir, config, nil); err != nil {
+		if err := sanitizer.DeterminizeJobs(subDir, config, nil, blocked); err != nil {
 			logrus.WithError(err).Fatal("Failed to determinize")
 		}
 	}
