+++ test/integration/cluster-init/update/expected/core-services/sanitize-prow-jobs/_clusters.yaml
@@ -1,0 +1,3 @@
+aws:
+  - build01
+  - build02
