+++ test/integration/repo-init/expected/core-services/sanitize-prow-jobs/_clusters.yaml
@@ -1,0 +1,1 @@
+aws:
