+++ test/integration/repo-init/input/core-services/sanitize-prow-jobs/_clusters.yaml
@@ -1,0 +1,1 @@
+aws:
